requirements.txt
aiogram==3.4.1
aiohttp==3.9.5
openai==1.30.1
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.1
motor==3.4.0
pymongo==4.6.3
apscheduler==3.10.4
prometheus_client==0.20.0
